phase one of implementing music playing
